Add a bastion host in azure
We add the first configuration of a host in azure. To add
a host there are two requirements: 
 * A hosted service, which we decided to declare it explicitly per host in terraform.
 * A SSH fingerprint.

The SSH certificate is associated to the hosted service created above, and needs to be uploaded to the console. So far we did not find any way to setup it in terraform. This is an issue, as currently you need to first run terraform once (which will create the service and fail due missing SSH key) upload the certificate manually, and run terraform again.

We will investigate alternative approaches.