Add provisioner to upload ssh key to azure
In order to [automate the process to generate and upload the 
SSH certificate](https://media-glass.es/2015/07/22/adding-a-ssh-key-to-azure/) 
to azure, we add a `local-exec` command with all
the required steps and that calls the azure client to perform the 
upload.

This adds a explicit dependency on the azure client.

The uploaded ssh certificate thumbprint is saved in a temporary file
called `ssh_thumbprint`, which terraform reads 
(added empty and ignored in the repo)