Add Azure endpoint ACLs to restrict SSH access
We implement [Azure's Endpoint ACLs](https://azure.microsoft.com/en-gb/documentation/articles/virtual-networks-acl/) to
restrict the access to the bastion SSH port to allow only GDS office IPs.

From the documentation, the ACLs rules will block everything but the traffic that is explicitelly allowed:

> **Permit** - When you add one or more "permit" ranges, you are denying all other ranges by default. Only packets from the permitted IP range will be able to communicate with the virtual machine endpoint.

Endpoints ACLs is not currently supported by terraform (=< 0.6.3), so we will use a `local-exec` provisioner in the `azure-instance` resource which will call the azure-cli command `azure vm endpoint acl-rule create ...`.

In order to be able to use the comma separated list of office IPs defined in `globals.tf`, we created a wrapper script which will split this list and add a sequence of rules, increasing their priority by one. This way it is easier to modify the list of allowed CIDR.