firewall: only local network for bastion to bosh
There's no need to access microbosh on it's external IP when it resides on the same network as bastion host. Only allow bastion to microbosh communication to happen on local network. No other communication to microbosh is allowed (not even from office: if you want to manipulate microbosh, do it via bastion host).