bastion host: use fixed IP
Use 10.0.0.4 fixed IP for bastion host. As we might want bosh to deploy something in the spare range of this subnet (say from 10.0.0.10 to 10.0.0.250), we fix the IP of bastion host to 10.0.0.4, the 1st available IP in AWS subnets (.1 is gateway, .2 DNS and .3 is also reserved). Specify bastion host VPC security group explicitly, using correct terraform argument.