Add script to create an azure storage account
Terraform does not support creating azure storage accounts, so
we need to use the `azure-cli` for the time being.

We provide a external script `azure/azure-create-storage-service.sh`
which can be called by `local-exec` provider.

This script creates the storage account associated to a given
service name (terraform can create those) and retrieves and
stores the account key in a file which we can read in terraform later.

If the account already exists, it just downloads the key.