Add an azure hosted service for storage accounts
The storage accounts, needed to be able to create storage resources in
azure by the BOSH cpi, are associated to a `azure_hosted_service`.

The `azure_hosted_service` is supported by terraform, but the account
not, for which we have the script `azure/azure-create-storage-service.sh`.

We add a dedicated service for storage with a `local_exec` provisioner
to create the storage account.

Eventually, if storage accounts are implemented in terraform, we can
move the logic to terraform itself.