Add script to create azure app service principal
BOSH Azure cpi requires a new client ID with a password created in the
`active directory` of the Azure account with a contributor role
to login in azure and create the different objects.

In order to create this credentials, we must:
 1. create an application with a url and a given password
 2. create a service principal associated to that application
 3. add it to the `Contributor` role

Terraform does not support this, so we created a script to implement this
logic which we will be able to call manually or from terraform.

This script creates the client id used in the Bosh manifest.

Note: The created user cannot be deleted from the commandline (at least
I did not figure how). To delete it: Azure console > Active Directory >
select the existing AD > Applications > Select "apps owned" > Click on
the given app > delete (icon in bottom bar)