Split azure_hosted_service in storage to reuse it
In BOSH we need a Azure resource group to use in all the created assets
(vms, disks, etc). The terraform `azure_hosted_service` resource creates a
Azure resource group.

We will creat a global `azure_hosted_service` and used to create all the
objects.

We change the storage account to be created within this new global
group.