Create cf network in azure with external command
BOSH azure cpi requires the used network to be created within the
resource group assigned to BOSH. But the [azure network terraform
resource](https://terraform.io/docs/providers/azure/r/virtual_network.html) does
not allow to specify the resource group.

Because that, we avoid using the terraform resource and we call the
azure command line directly, to create the network and subnet in the
right resource group for BOSH.

We will need to solve how we will connect this two networks.