Add necessary endpoint for public ip of azure bosh
In order to be able to connect to the service of microbosh on Azure via
the Public IP, we need to define the required ports as endpoints.

We start opening only SSH, and probably bosh-init would be able to
provision using the `ssh_tunnel` defined later.

NOTE: This is opening the ports to the public, we need to fix it.