Add endpoint to access bosh agent via public IP
As the bastion host where we run bosh-init is currently running in a different 
network than the bosh machine, we need to map the port 6868, used by the bosh agent
in the stem cell, so bosh-init can connect to it and setup the microbosh machine.

The `cloud_provider.ssh_tunnel` does not work for this case :(

An alternative solution could be point the variable `cloud_provider.mbus` to localhost:

```
cloud_provider:
  mbus: https://mbus-user:mbus-password@127.0.0.1:6868
```

and then manually create a tunnel with SSH with `ssh vcap@23.97.216.207  -i .ssh/id_rsa -L 6868:localhost:6868`

But given that the current situation (bastion and bosh in different networks) is in general temporary and suboptimal, we will just configure the port in the public ip for the time being.