Open ports for bosh-deployed VMs
The bosh deployed VMs need to communicate with the microbosh on a
number of different TCP and UDP ports in order to get their
configuration information, fetch packages, resolve DNS and perform
logging. Due to the weakness in GCE's network implementation, this
essentially involves either opening those ports to the world, or
running all outbound traffic through a SPOF address translation box
so that outbound traffic comes from a known-good IP address.