Move upload of manifest out of instance provision
The provisioners associated to the instances will only update
the files if the instance is reprovision. This means that
we will need to reprovision the bastion host every time we want
to change the template.

To avoid this, we will create the files using a `local-exec` provisioner
in the terraform template and upload them using ssh in the Makefile.

One restriction of this is that the template must not include a single
quote, as it might interfere with the local-exec of the template.