GCE bastion: Increase disk to 100GB
Lots of disk space is consumed by bosh to download and manipulate stemcells and manifests. For example cf210 release requires around 13GB of disk space. We might want to keep multiple versions of stemcells and releases locally on the disk. Bosh client also keeps it's beefy local cache on the disk.