GCE: make default subnet same as in AWS
The original 10.10.0.0/24 was chosen just because there was some example that had this subnet. There's no neason we should have (or need to have) different subnets in GCE and AWS. Change the subnet to 10.0.0.0/24 as it is in AWS. This can be later unified and pulled out into global variables.