Use public EIP for AWS microbosh to access inet
The microbosh machine needs to be able to reach the internet, to
download binaries and connect to the AWS API.

In order to do that, we can either setup a NAT host to route (which is a SPOF)
or assign a public IP to the host. We apply here the second approach.

The security groups associated to the director will block any incomming
traffic.

This commit partially reverts d2986a5ff9d29fdaa4b8b5944431c69b74651726