GCE manifest: render using erb
GCE cf210 manifest comes as ERB. This is convenient, as we can use erb to render it. In terraform we just export a small erb snippet that defines variables with values from terraform (deployment name, network name and haproxy IP). Remove the old template as we don't use it anymore.