GCE destroy: Delete the microbosh route
When decomissioning environment, we have to delete the route we created for re-routing microbosh external traffic internally. Otherwise terraform will not be able to destroy the network, due to being blocked by existing route - which was not created by TF. Add new makefile target and a script to delete the route. This can be further improved when we do future work on improving route handling.