GCE firewall: only internal access to microbosh
Close any external access to the microbosh. Instances now access it internally, thanks to re-routing of the traffic to the bosh external IP. Current security groups still allow all-to-all communication between CF instances, this should be tightened up later. The haproxy instance (CF entry point) still has 80, 443 and 4443 open to the world, as it is being accessed by ephemeral external IPs of the other CF components. This is because CF components access each other for some operations (e.g. accessing UAA when logging in) on the official deployment URL, which points to external haproxy IP.