GCE firewall: open internal all to all
Opening GCE firewall for internal communication for all to all traffic on UDP and TCP. Reason is that we're finding one-by-one various CF functionalities disabled by firewall restrictions. It's always a tedious process to find out which ports exactly need to be open. Opening all-to-all now, so that we have a fully functional platform. We'll be restricting ports later, possibly also using different tags for different host. Keeping list of known ports to be open in the tf file comments.