Use the private networks in AWS manifest stubs
Use the two private networks in different AZ in the cf template.

Note: Currently we are not sure why originally cf1 was map to 'private' and cf2 was map to 'public'. But as it seems that the AWS template intentions is have networks with the same setup in different AZ, we will use 'private' both cases.