Remove nat, cf_route and default AWS securitygroup
We have decided not to setup bastion as a NAT host **for the time
being**, and instead use first public IPs for all the VMs, later setup
nat in one go.

Because that, we do not need all the security groups copied from the
tsuru configuration. This simplifies the configuration as we only keep
these security groups: bastion, director, bosh_vms and web.