Open AWS LB endpoint to the world
Several internal services in CF need to connect to the public endpoint
of the platform to access services like loggregator, uaa, etc.

As we decided to do not setup a NAT, all VMs will be routed via their
random public IP, so we won't not able to restrict access from the
office and those IPs.

As a short term solution, we just open it to the world and we will setup
NAT and restrict it later.