Use official spiff templates for AWS CF manifest
To use [the official templates from the cf-release](https://github.com/cloudfoundry/cf-release/tree/master/templates) we will upload a stub which populates the required variables, and use [spiff to render the configuration](https://github.com/cloudfoundry/cf-release/blob/master/generate_deployment_manifest).

We need to remove the old manifest and upload the new stub. The manifest is 
rendered in the bastion host as part of the `provision.sh` script, so we need
to install the `spiff` command.

We commit an empty stub and will comment each change as needed, so the configuration is 
not functional.