Use public ELB to load balance routers
Add a ELB to load balance the CF routers. The ports will be 80, 443 and 4443 (used by default by loggregator web sockets).

BOSH can automatically add member to a ELB when deploying new VMS, but for that it needs credentials which we pass in the template. It is pending to test with a AMI policy to prevent this.

We open the port 4443 because the [CF templates setup loggregator
endpoint to this port](https://github.com/cloudfoundry/cf-release/issues/746#issuecomment-126497276)
to be able to define different configurations
for the normal traffic and the loggregator, as ELB HTTPS does not handle
websockets. Although we still use `tcp` in 443 to avoid having to create
the certificate on AWS, we will keep using 4443 for loggregator to be
more aligned to the upstream config.

NOTE: We currently add a security group which allows access from everywhere, to allow
all the instances which will get random public IPs access the CF API endpoint. Setting
security groups does not seem to work with public IPs. This needs to be sort out in the future
by using a NAT gateway or enforcing internal communication.