Basic configuration for cf-infrastructure-gce.yml
Add the basic entries for the GCE infrastructure, inspired in the AWS config.
This includes:
 * stemcell
 * network config
 * compilation properties
 * loggregator and uaa config 

Note that for loggregator we set port 443 instead of 4443 as in AWS. The reason is 
that the GCE load balancers do not support set a different port in the backend than 
the one in the Load Balancer IP, and the gorouters only listen on 443. 

This adds the restriction of using only TCP load balancing to support websockets.