Configure auto registering of routers in GCE LB
BOSH Google CPI support self registering of jobs in Google
load balancers. For that, new networks with the 
`cloud_properties.target_pool` property must be defined 
and assigned to the jobs.

We create two networks, `route1` and `route2` for consistency with
the existing networks `cf1` and `cf2`, and override the network
settings in the router jobs and resource pools.

This change requires a [patched version of cf-release](https://github.com/alphagov/cf-release/tree/cf_jobs_without_static_ips_dependencies_v215).
    
[Related PR](https://github.com/cloudfoundry/cf-release/pull/791)