GCE: disable resurrector
In GCE, VMs get new IP after resurrecting. This breaks things, as local jobs on the VM still think they have old IP and try to run on old interfaces (causing them to fail to start) or even register the old IPs in the router, making the services unaccessible. This can be enabled again once we implement #103979660.