Use a sourced .sh file for terraform output vars
Parsing the terraform_outputs Yaml file in bash is tricky and error
prone. It can break if the Yaml file has some complexity or a
variable with multiple lines.

Instead, we generate from Makefile a valid shell include file with the
required variables, which can be included by `provision.sh`