Update stemcell to 3074 and BOSH release to 207
Updating stemcell versions. In the case of the GCE stemcell, this version
fixes the [ssh issue #103408576](https://www.pivotaltracker.com/story/show/103408576).

This stemcell is based on AWS 3074, and we want to use the same version
on both platforms, so we also update AWS. Finally, it is convenient
align the stemcell with the latest BOSH release (v207 at the moment).

Note that we update the url for GCE to use our own custom stemcell url
in a different bucket in google storage.