Adding networks to microbosh job
The only part that differs is the static IP for the private network,
which is not supported on GCE, so it must be overwritten in the
GCE stub with a value `null`.

The default IPs will be the bosh private or public IP in each case,
as a list as `network.static_ips` is a list.

We create two structures `meta.default_bosh_private_static_ips` and
`meta.default_bosh_public_static_ips` which are lists of IPs in YAML
and allow us to optionally override all the list with null using the
idiom: `static_ips: (( merge || meta.default_bosh_private_static_ips ))`