Use bundle exec command to run ruby based scripts
Use bundle exec to call any of the ruby commands in provision.sh. This
includes the `bosh_cli` and a pair of utility scripts.

In order to be independent of the current working path, we set the
variable `BUNDLE_GEMFILE` to point to the right `Gemfile`.

This is actually the only way to pass the `Gemfile` path to
`bundle exec` in this version of ruby.