Check for missing packages with outdated index
If the dpkg database index is not updated, the `dpkg-query` command
might fail with `dpkg-query: no packages found matching libssl-dev` if
the package is not found.

To workaround that issue, we implement a more smart check. It will list
all the packages and redirect all the `stderr` output to `stdout`.
If  there is any line in this output which does not include
`ok installed`  it means we found a missing package (uninstalled or not
found).