Set dea_next MTU property
We are setting the MTU to be 1460 in the warden containers to match
the MTU on the runner VMs.

We also need to reference a new branch based on the
cf_jobs_without_static_ips_dependencies_v215 branch that allows us to
set the dea_next MTU property as a mergeable parameter.

Reference:

* https://github.com/alphagov/cf-release/pull/1
* https://github.com/alphagov/cf-release/tree/cf_jobs_without_static_ips_dependencies_v215_103419194_with_dea_next_mtu