Use specific IaaS stub after common for CF manifest
When generating the CF manifest, we put the specific IaaS file as the
last one in spiff.

We need this change to be able to specify common configuration for
databases later.

As a drawback, we need to duplicate the definition of the name of the
deployment as we need it in the specific gce stub.