Use BOSH deployed postgres instead of RDS on AWS
Stop using RDS but a VM running postgres for CCDB and UAADB, 
in the same way as we do in GCE. 

This will make both environments more similar, using the common
denominator, but also reduce deployment steps as postgres will be
deployed by BOSH in parallel with the rest of components.

In this commit we move all the definitions related to the databases
to the shared stub. 

The only non common configuration is the postgres address: We use
BOSH powerdns on GCE and static IPs on AWS. 

To enable postgres on AWS, we increase the number of instances in the
AWS stub. We also specify the static IP for AWS. We chose the spiff
generated `static_ip(7)` [as it is defined in the official cf-release
for warden.](https://github.com/cloudfoundry/cf-release/blob/master/templates/cf-infrastructure-warden.yml#L454).