Pass cf admin credentials to the smoke test script
The password for the admin user has been changed to a secure password
stored in our password store. We need to pass the new credentials
from `paas-pass` for the admin user to be able to run the smoke tests.

This depends on the right credentials stored in: `paas-pass cloudfoundry/cf_admin_password`