cf-jobs: deploy all in parallel
The only job that defines serial deployment is consul.
But CF is expecting (or is promised to be able to deal with)
all jobs deploying in parallel. Monit simply keeps restarting
the processes until all dependencies are running. This is the
fastest way to deploy and also most robust. With serial being
set, jobs are split into 2 batches to deploy. If a job in 1st
batch depends on job in 2nd, it fails at which point BOSH stops
deployment. Because if spiffy nature of spiff, it's hard to tell
where our jobs would end up in the final manifest -if before or
after job with serial updates defined. Currently only consul has
serial updates and it was moreover defined as 1st job. But our
spiff adds some our jobs before it anyway. Hence disabling serial.