Increase the ELB idle timeout to 300 seconds.
The AWS elstic load balancer will timeout connections after a number
of seconds of inactivity. By default this is set to 60 seconds, but this
was causing problems with the deployment of the postgresql-service-broker
when pushing the app to a freshly deployed CF.

After investigation, it was found that there is a long period of inactivity
between the CF CLI client and the API server whilst the DEA downloaded and
unpacked the buildpacks required by the psql broker. It appears that we need
to allow at least 157 seconds for these buildpack steps to complete the first
time.

Increasing the AWS ELB idle_timeout up to 300 seconds allows for all of the
buildpack steps to complete, and allows the psql broker to deploy without
error.