Fork graphite-nozzle; prefix statsd metric names
Use our fork of graphite-nozzle which, when combined with the `PREFIX_JOB`
environment variable, will prefix each metric name with the job that it came
from (e.g. `runner_z1.0`). This will allow us to debug issues like comparing
doppler send and receive counts on different VMs.

I'm going to pull request this upstream, but we're not sure at this stage
whether it will be accepted. So I'm using a fork for the timebeing. We
should review this if we rebuild for beta.

In order to switch to the fork I have added a new remote, ensure that we
have fetched from it, and then checkout detached HEAD against the latest
commit on that remote branch. The `git reset` is needed because we modify
`Godeps.json` and can't checkout a different commit with a dirty tree. The
`git clean` is a safety measure, which we don't strictly need because
`manifest.yml` and `Procfile` aren't checked in, but we might have other
files that prevent changing branches.