Stub elasticsearch IP from bosh output in GCE
Currently the errand that deploys kibana-for-cloudfoundry tries to setup a security-group on CF based to allow access to elasticsearch based on the IP defined in the manifest. There is no way to 
specify an IP range and DNS aliases are not valid in this case.

On GCE This is an issue, as GCE CPU does not support static IPs, 
but the VMs get IPs dynamically allocated.

The solution requires either do a pull request or deploy kibana out 
of the errand, but as a **sort term** solution we will workaround 
this by:

 1. Deploy logsearch (Elasticsearch+Lostash as usual)
 2. Query bosh to get the allocated IP.
 3. Generate a spiff stub with the IP for eleasticsearch.
 4. Run the errand with the new config.