Use a helper script to get the bastion IP or host
When connecting to the bastion, we want to keep using the ip retrived
from terraform if it is available, but if not, try to use DNS.

This will allow us to connect to any environment by name, even if we
don't  have the terraform state file. We don't use the DNS when
the terraform file is available because just after provisioning the DNS
entry might have not propagated yet or a negative entry can be cached.

Use a external helper to encapsulate the logic to get the bastion IP if the
terraform file is present and populated, or generate the DNS name instead.