Redirect HTTP traffic to HTTPS on the routers
gorouter has no capability to redirect HTTP traffic to HTTPS. In order
to work around this we install an nginx release on the routers listening
on port 80, while reconfiguring gorouter to listen on port 8090 (out of
the way). It can't be reconfigured to disable listening on a plain HTTP
port.

The nginx release is from cf-community:
https://github.com/cloudfoundry-community/nginx-release

CF validates that an app has correctly started by polling it via the
full HTTP routing pipeline - this breaks when we're redirecting port 80,
as it doesn't expect to get a 301 response code. By specifying
````external_protocol```` as HTTPS we forcing CF to use 443 for health
checks.

This commit also adds a deployment ````https_redirect.yml```` template
as nginx needs to be colocated on the same machine as the routers, which
are the ELB targets.

This commit depends on PR #2 in our cf-release fork in order to allow
overriding the gorouter port:
https://github.com/alphagov/cf-release/pull/2

Initially we investigated a cf community HAproxy release that has
built-in functionality for redirecting all HTTP requests to HTTPS.
After a substantial amount of effort we completely failed to get it to
work as the final release appeared to be malformed, and the development
release installed its templates incorrectly as "haproxy". Additionally
the misnamed templates lead to concerns about conflicts with the haproxy
release packaged by cf-release. Nginx was simple to install and
configure, and is guaranteed not to cause problems with other
components.

https://github.com/cloudfoundry-community/cf-haproxy-boshrelease