Create an AWS ELB to expose graphite to the office
Add a new ELB to expose graphite and grafana to the office. 
We also setup the security group for the graphite and 
grafana ports, the required DNS alias with the same format and 
naming than tsuru (`<env>-grafana.cf.paas.alphagov.co.uk`) and
the required terraform output to use the ELB id in the CF manifest.