Graphite manifest allows LB registering on AWS
The Bosh CPI in AWS and GCE support self registering of the VMs in
the loadbalancer by using `cloud_properties`: in the case of AWS
in the instance, in GCE in the network).

Because that, we need to define a special `resource_pool` for the
graphite server, copying the definition of `large_z1`. The IaaS
specific config should be added in a different stub, in this case
for AWS.