Create an GCE LB to expose graphite to the office
Add a new LB (address+pool) on GCE to expose graphite and grafana to
the office and jenkins.

We also setup the firewall for the graphite and grafana ports, the
required DNS alias with the same format and naming than tsuru
(`<env>-grafana.cf.2paas.alphagov.co.uk`) and the required terraform
output to use the pool name in the CF manifest.