Graphite manifest allows LB registering on AWS
GCE BOSH CPI support self registering of the VMs in a LB pool by
adding some `cloud_properties` in the network definition.

We need to define an additional `resource_pool` and `network` for
graphite copying the `large_z1` definition.

In GCE, as there are no static IPs we rely on the BOSH dynamic DNS,
which refers the network name of the specific job/vm, so it needs to
be updated with the new network name: `graphite1`