Logstash drops bogus field in router-api logs
The field data is a hash that fails inserting in the ElasticSearch
cluster with this error:

```
{:timestamp=>"2015-11-12T14:06:53.252000+0000", :message=>"failed action with response of 400, dropping action: [\"index\", {:_id=>nil, :_index=>\"logstash-2015.11.12\", :_type=>\"relp_cf\", :_routing=>nil},
...
```

Droping all the parsed `data` field fixes the issue. We suspect that it
was caused by uppercase keys in the hash, but lowercasing did not fix
the issue, so the root cause is unknown.

The full raw message will stay in `@message`