Use newly split out dev/trial SSL certificates
This adds a new "aws-trial" deployment target that uses the new trial
SSL certificates that have been added to paas-pass. On deployment the
new secrets files are decrypted, uploaded to the bastion, and then
merged into the manifest at generation time.

The old "aws" target has also been changed to use the new location of
the dev SSL certificates and perform the same actions.