Use make functions instead of directives
In Makefile we intend to check the value of DEPLOY_ENV and WEB_ACCESS_CIDRS in
the task. But ifndef is actually a directive that is evaluated always,
despite if the target is called or not.

Because that, Make will always fail if DEPLOY_ENV is not set even when
running tasks that do not require that variable.

This also prevents us to do syntax checking with `make -n Makefile`

This commit fixes the issue by using conditions rather than directives.