Working implementation of Dockerised BOSH tools
Running "docker build ." will generate a container that has bosh and all
it's dependencies in it, along with other tools that we require for
provisioning. You can bosh-init from within the container, along with
all other bastion tasks.

To use the container you'll want to mount a directory containing all the
files required by bosh when provisioning that are outputted from the
Make process, e.g.:

  docker run -ti -w /root/ --rm=true -v
/Users/jonty/gds/cf-terraform/:/root/ 130f2b52822c
/home/root/scripts/provision.sh aws

The current provision script will fail when running directly from
cf-terraform as the directory structure that we currently create on the
bastion is slightly different. If we wish to adopt this approach fully
then we need to refactor our layout and provisioning scripts to be
identical, and ideally the makefile should be able to output the spiffed
manifest layout locally without copying to the bastion.

We also copy over the Gemfile (and .lock) to this directory as docker
will not allow you to add files outside the current working directory,
and does not follow symlinks. We should change the repository layout to
ensure we only have one copy of these files.