Update combined dashboard
Update graphs shown on the combined dashboard. Over the past week we
have collected a set of comments from our team members with various
suggestions on how to improve graphs - change scales, change description
of metrics, add or remove some metrics, etc. This is second iteration
of this dashboard and we expect to update it again in the future, as
the usage pattern of our environment changes (more users, different usage
patterns (e.g. CI)) - which is expected to be quite soon.

Added:

* Router load
* Disk IO state for the runners
* The DEA names were being truncated so we shortened the names
* Add Changed Warden Response time to Warden MS to make sure the values fitted on one line
* Cloud Controller Requests has been moved from App Metrics to Cloud Controller Metrics graph
* Removed Duplicate DEA Metrics `dea_registry_stopping` metric in the App Metrics graph and replaced them with:
 * `dea_registry_crashed`
 * `dea_logging_agent.totalApps`
* Set Y axis to start at zero for the following graphs
 * DEA stats
 * CPU load
 * DEA memory used
 * DEA disk
* Upper limit on Y axis to sane values e.g. 100% or maximum value for a graph
* Title fixes