Use SSD for graphite's persistent disk.
The graphite machine was using magnetic disks for its persistent disk.
This couldn't keep up with graphite's write load.

In order to specify cloud_properties for a persistent disk it's
necessary to use disk pools[1], so this PR adds a disk pool for the
graphite persistent disks, and updates the job to use it.

This depends on https://github.com/alphagov/cf-release/pull/3 being
merged first.

[1] https://bosh.io/docs/persistent-disks.html