Delete innactive metrics in statsd
Enable the option `deleteIdleStats` in statsd to drop innactive metrics
for counters, sets, gauges, or timers. This will save space as statsd
will stop sending metrics of old metrics if the resource is not present
anymore. This is important in a dynamic environment like CF, specially
for app metrics.

There is a side effect: if a metrics is not send for a whie, the
values create "holes" in the graphs with `null` values.

This can be alleviated by using graphite functions transformNull for
counters and keepLastValue for gauges .