Deploy kibana proxy to CF
Deploy a basic auth enabled proxy for kibana into CF. This is in order
to make kibana more easily accessible (otherwise you have to make SSH
tunnels via bastion and add kibana into your hosts file). The basic
auth prevents unauthorized access. The deploy script also creates the
necessary security group to let the proxy reach kibana server.

This is deployed as a replacement for the UAA enabled kibana that has
disk read issues. Once we get that app working we can revert adding this
proxy and the kibana on the logsearch servers instead. The code of the
deploy script could be DRYed with other 2 CF app deployment scripts
(graphite nozzle and psql), but this is just a temporary workaround on
a legacy build, so no further effort to improve things was spent.