Increase DEA resource limits
The limits were set to 10G disk and 4G (x3 overprovision) memory. We
run r3.large DEAs with 15G of ram and 100G of disk, out of which 80G
is available. Increasing resource limits conservatively to 60G disk and
8G (x3 default oveprovision) memory. The apps usually consume 50-70% of
memory and even less of disk allowance, so these limits are safe.