Merge in terraform outputs.
This adds the terraform outputs into the spruce merge command. There's
currently nothing taking anything from these yet though, so there will
be no difference in the output.

This also sets up the tests to use a fixture version of these outputs
which contains the subset of the outputs that are currently used in the
deployment.