Refactor bosh manifest from spiff to spruce
Merging in spruce is done a little differently.  From the docs:

  Files to merge are listed in-order on the command line. The first file
  serves as the base to the file structure, and subsequent files are
  merged on top, adding when keys are new, replacing when keys exist. This
  differs slightly in mentality from spiff, but hopefully the results are
  more predictable.

This means that it's not necessary to list hash keys at the top-level in
order for them to appear in the output.

We also remove '(( merge || <literal> ))' operations

Due to the way spruce merges these, these can be literal values, spruce
will automatically overwrite them if a subsequent file includes a value.

[1] https://github.com/geofffranks/spruce#merging-rules