Logsearch: set ELS replicas to 0
We run logsearch on a single host, which also hosts elasticsearch.
Because of this, having shard replicas is pointless, as there is no
other elasticsearch server to replicate to. Yet when you set replicas
to 1, replicas will still be created. This means that the cluster
status will be yellow (as replicas are not replicated) and the mem/
disk space consumption would be larger, as we have 2 copies of data.

Set replicas to 0 to make server faster (less data to manage by
removing duplicate data) and cluster status green. Except kibana 4.1.1
still hardcodes creation of 2 replicas. Kibana 4.1.2 should support
user configured replication settings.