Upgrade bosh releases to get iam profile support
The releases for BOSH and the AWS CPI need to be upgraded to support
the configuration of IAM profiles.

In the case of BOSH, the next version (v207 => v208) supports this
configuration, and we keep the changes to a minimum.

In the case of AWS CPI, we upgrade to the latest version to benefit of
some additional bugfix.

The new AWS CPI release renames the cpi job to `aws_cpi`, so we need to
change some configuration in the bosh deployment to keep everything
working. For instance: `director.cpi_job` which renders the path to the
CPI driver.