Use IAM profiles for S3 access from API
Use IAM profiles for S3 access from the API to the cf assets, including:
droplets, packages, build packs and resources.

For that we need the changes in https://github.com/alphagov/cf-release/pull/4
which allow us to specify an alternative resource_pool for the api nodes
which includes the settings for the IAM profile.

We override the resource pool in the api_z1 and api_z2 jobs, and pass
the option `use_iam_profile` to AWS hash, which will enable IAM profile
in the fog library.

Note that this AWS settings hash is passed without any change directly
to the controller configuration, which passes it directly to the Fog
library contructor. This means that any valid config for the fog library
is valid here.