Remove postgreSQL service purging scripts
Even though the purging scripts are idempotent, meaning they check if
the service has been removed already and do nothing if it has, there is
no need to keep running them after the service was removed.

Remove these scripts, references to them and the dummy postgres yml
files used by spiff. The dummies deployed already will not get removed,
but this is OK as the files are not adding anything to the main
manifest already.