Adjustment of network configuration
Adjust ranges from default CF manifest to match those in our VPCs.
Remove `cf` security group usage, as that SG doesn't exist. Instead
default BOSH security group was used (this is specified in the
microbosh manifest). Remove this declaration, so that it is clear that
the real security group definition is elsewhere:
https://github.com/alphagov/cf-terraform/blob/master/bosh-manifest/deployments/aws/090-bosh-aws-stub.yml#L6