Add a default type for timestamp property
Our current logstash configuration simply sends all the parsed json
from all the CF components to Elasticsearch. But different applications
might send different types for the same fields, which can cause issues
in ES metatada when the indexes are rotated everyday, as conflicting
types might be detected.

Specifically, we had issues with the mapping of repl_cf.log_level and
repl_cf.timestamp. Sometimes a long or double value is sent as first
message just after rotating the index, causing ES to automatically map
the field to that type and reject any other message which send strings
instead.

The short term solution is setup a default field mapping and forcet the
type for repl_cf.log_level and repl_cf.timestamp to be string.

Later we will update the parsing rules in logstash, so these values
are filtered and mapped properly.