Drop data field for HM9000 messages from CC
Some messages from HM9000 including fields with health of the apps
are parsed as json and ingested into ES.

As smoke tests deploys apps with a UUID every some minutes,
thousands of fields are created, which crashes kibana when
queries for the existing fields.

We don't really use this log messages, so we can simply drop the
data field as a short term solution.